+++ docs/reference/changelog.md
@@ -2,7 +2,7 @@
 
 ## Versions
 
-- [Webots R2024](changelog-r2024.md)
+- [Webots R2025](changelog-r2025.md)
 - [Webots R2023](changelog-r2023.md)
 - [Webots R2022](changelog-r2022.md)
 - [Webots R2021](changelog-r2021.md)